+++ CHANGELOG.md
@@ -2,10 +2,24 @@
 
 Semua Perubahan Dan Penambahan Feature Ada Di Sini.
 
+## 2.0.0 - 15/02/2023
+
+### Scan Folder (New)
+
+- Fitur ini adalah pembaruan dari fitur Scan Folder yang lama, yang dimana terkadang membaca semua nama file dan menyimpan hasil nama nya dengan acak, dan kita harus menyusun nya kembali, dengan fitur ini program akan membuat folder baru lalu mengcopy dan menambahkan angka pada setiap nama file, dan hasil nama file nya juga akan berurutan.
+
+### Add Features
+- Menambahkan console menu Scan Folder (New)
+
+### Change Features
+- Merubah nomor console menu untuk Scan Folder dari 1 menjadi 2
+- Merubah nomor console menu untuk keluar dari 2 menjadi 3
+
+---------------------------------------------------------------------------
 ## 1.0.0 Release - 10/02/2023
 
 Versi Awal
 
-### Add Feature
+### Add Features
 - Menambahkan Fitur Scan Folder Tree, dengan format yang sudah di tentukan.
 - Menambahkan Fitur Keluar Dari Console.